+++ src/slave/containerizer/docker.hpp
@@ -112,6 +112,9 @@ public:
   process::Future<ContainerStatus> status(
       const ContainerID& containerId) override;
 
+  process::Future<std::pair<std::string, std::string>> logs(
+      const ContainerID& containerId) override;
+
   process::Future<Option<mesos::slave::ContainerTermination>> wait(
       const ContainerID& containerId) override;
 
@@ -166,6 +169,9 @@ public:
       const ContainerID& containerId);
 
   virtual process::Future<ContainerStatus> status(
+      const ContainerID& containerId);
+
+  virtual process::Future<std::pair<std::string, std::string>> logs(
       const ContainerID& containerId);
 
   virtual process::Future<Option<mesos::slave::ContainerTermination>> wait(