+++ src/slave/http.hpp
@@ -182,6 +182,11 @@ private:
       ContentType acceptType,
       const Option<process::http::authentication::Principal>& principal) const;
 
+  process::Future<process::http::Response> readLog(
+      const mesos::agent::Call& call,
+      ContentType acceptType,
+      const Option<process::http::authentication::Principal>& principal) const;
+
   process::Future<process::http::Response> getFrameworks(
       const mesos::agent::Call& call,
       ContentType acceptType,